fix(signature-collection): renaming templates to presidential specific
