+++ libs/application/template-api-modules/src/lib/modules/templates/signature-collection/signature-list-signing/signature-list-signing.service.ts
@@ -7,7 +7,7 @@ import {
   SignatureCollectionClientService,
 } from '@island.is/clients/signature-collection'
 import { TemplateApiError } from '@island.is/nest/problem'
-import { errorMessages } from '@island.is/application/templates/signature-collection/signature-list-signing'
+import { errorMessages } from '@island.is/application/templates/signature-collection/presidential-list-signing'
 import { ProviderErrorReason } from '@island.is/shared/problem'
 
 @Injectable()
@@ -15,7 +15,7 @@ export class SignatureListSigningService extends BaseTemplateApiService {
   constructor(
     private signatureCollectionClientService: SignatureCollectionClientService,
   ) {
-    super(ApplicationTypes.SIGNATURE_LIST_SIGNING)
+    super(ApplicationTypes.PRESIDENTIAL_LIST_SIGNING)
   }
 
   async signList({ auth, application }: TemplateApiModuleActionProps) {
