+++ libs/application/templates/signature-collection/presidential-list-creation/.babelrc
