+++ libs/application/templates/signature-collection/presidential-list-creation/README.md
@@ -1,0 +1,7 @@
+# application-templates-presidential-list-creation
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test application-templates-presidential-list-creation` to execute the unit tests via [Jest](https://jestjs.io).
