+++ libs/application/templates/signature-collection/presidential-list-creation/src/dataProviders/index.ts
