+++ libs/application/templates/signature-collection/presidential-list-creation/src/fields/ListCreated/index.tsx
