+++ libs/application/templates/signature-collection/presidential-list-creation/src/fields/index.ts
