+++ libs/application/templates/signature-collection/presidential-list-creation/src/forms/Draft.ts
@@ -19,7 +19,7 @@ import { formatPhone } from '../lib/utils'
 import format from 'date-fns/format'
 
 export const Draft: Form = buildForm({
-  id: 'SignatureListCreationDraft',
+  id: 'PresidentialListCreationDraft',
   title: '',
   mode: FormModes.DRAFT,
   renderLastScreenButton: true,
