+++ libs/application/templates/signature-collection/presidential-list-creation/src/index.ts
