+++ libs/application/templates/signature-collection/presidential-list-creation/src/lib/constants.ts
