+++ libs/application/templates/signature-collection/presidential-list-creation/src/lib/dataSchema.ts
