+++ libs/application/templates/signature-collection/presidential-list-creation/tsconfig.lib.json
