+++ libs/application/templates/signature-collection/presidential-list-signing/.babelrc
