+++ libs/application/templates/signature-collection/presidential-list-signing/README.md
@@ -1,0 +1,7 @@
+# application-templates-presidential-list-signing
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test application-templates-presidential-list-signing` to execute the unit tests via [Jest](https://jestjs.io).
