+++ libs/application/templates/signature-collection/presidential-list-signing/project.json
@@ -1,0 +1,34 @@
+{
+  "name": "application-templates-presidential-list-signing",
+  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/application/templates/signature-collection/presidential-list-signing/src",
+  "projectType": "library",
+  "tags": ["scope:application-system", "lib:application-system"],
+  "targets": {
+    "lint": {
+      "executor": "@nx/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": [
+          "libs/application/templates/signature-collection/presidential-list-signing/**/*.{ts,tsx,js,jsx}"
+        ]
+      }
+    },
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": [
+        "{workspaceRoot}/coverage/libs/application/templates/signature-collection/presidential-list-signing"
+      ],
+      "options": {
+        "jestConfig": "libs/application/templates/signature-collection/presidential-list-signing/jest.config.ts",
+        "passWithNoTests": true
+      }
+    },
+    "extract-strings": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "yarn ts-node -P libs/localization/tsconfig.lib.json libs/localization/scripts/extract 'libs/application/templates/signature-collection/presidential-list-signing/src/**/*.{js,ts,tsx}'"
+      }
+    }
+  }
+}
