+++ libs/application/templates/signature-collection/presidential-list-signing/src/dataProviders/index.ts
