+++ libs/application/templates/signature-collection/presidential-list-signing/src/fields/ListSigned/index.tsx
