+++ libs/application/templates/signature-collection/presidential-list-signing/src/forms/Draft.ts
