+++ libs/application/templates/signature-collection/presidential-list-signing/src/index.ts
