+++ libs/application/templates/signature-collection/presidential-list-signing/src/lib/errors.ts
