+++ libs/application/templates/signature-collection/presidential-list-signing/tsconfig.json
