--- libs/application/templates/signature-collection/signature-list-creation/README.md
@@ -1,7 +1,0 @@
-# application-templates-signature-list-creation
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test application-templates-signature-list-creation` to execute the unit tests via [Jest](https://jestjs.io).
